Move FiltersButton component from containers to components folder - it'll be reused
