+++ frontend/src/containers/map/content/details/tables/global-regional/useColumns.tsx
@@ -4,8 +4,8 @@ import Link from 'next/link';
 
 import { ColumnDef } from '@tanstack/react-table';
 
+import FiltersButton from '@/components/filters-button';
 import { PAGES } from '@/constants/pages';
-import FiltersButton from '@/containers/map/content/details/table/filters-button';
 import HeaderItem from '@/containers/map/content/details/table/header-item';
 import { cellFormatter } from '@/containers/map/content/details/table/helpers';
 import SortingButton from '@/containers/map/content/details/table/sorting-button';
